Reorder abilities.cpp
1) Renames unit_ability_list -> active_ability_list
           unit_ability -> active_ability
   Becasue oftherwise it would be really unclear what the
   difference between unit_ability and unit_ability_t is.
   since the first deals with active abilities, i renamed it,

2) moved active_ability_list & unit_ability::effect to a new file
   now ability.cpp contains only, function implementations of
   unit and attack_type, this is because it was rather hard to
   get an overview of that file since it contained function
   definition of many header files.
