+++ projectfiles/CodeBlocks/wesnoth.cbp
@@ -1206,6 +1206,8 @@
 		<Unit filename="../../src/units/abilities.hpp" />
 		<Unit filename="../../src/units/ability.cpp" />
 		<Unit filename="../../src/units/ability.hpp" />
+		<Unit filename="../../src/units/active_ability_list.cpp" />
+		<Unit filename="../../src/units/active_ability_list.hpp" />
 		<Unit filename="../../src/units/animation.cpp" />
 		<Unit filename="../../src/units/animation.hpp" />
 		<Unit filename="../../src/units/animation_component.cpp" />
