+++ source_lists/wesnoth
@@ -374,6 +374,7 @@ tls_root_store.cpp
 tod_manager.cpp
 units/abilities.cpp
 units/ability.cpp
+units/active_ability_list.cpp
 units/animation.cpp
 units/animation_component.cpp
 units/attack_type.cpp
