+++ src/units/abilities.hpp
@@ -18,42 +18,3 @@
 #include "map/location.hpp"
 #include "units/ptr.hpp"
 
-class unit_ability_list;
-namespace unit_abilities
-{
-bool filter_base_matches(const config& cfg, int def);
-
-enum value_modifier {NOT_USED,SET,ADD,MUL,DIV};
-
-struct individual_effect
-{
-	individual_effect() : type(NOT_USED), value(0), ability(nullptr),
-		loc(map_location::null_location()) {}
-	void set(value_modifier t, int val, const config *abil,const map_location &l);
-	value_modifier type;
-	int value;
-	const config *ability;
-	map_location loc;
-};
-
-class effect
-{
-	public:
-		effect(const unit_ability_list& list, int def, const_attack_ptr attacker = const_attack_ptr(), bool is_cumulable = false);
-		// Provide read-only access to the effect list:
-		typedef std::vector<individual_effect>::const_iterator iterator;
-		typedef std::vector<individual_effect>::const_iterator const_iterator;
-
-		int get_composite_value() const
-		{ return composite_value_; }
-		const_iterator begin() const
-		{ return effect_list_.begin(); }
-		const_iterator end() const
-		{ return effect_list_.end(); }
-	private:
-		std::vector<individual_effect> effect_list_;
-		int composite_value_;
-};
-
-
-}
