+++ src/units/active_ability_list.cpp
@@ -1,0 +1,352 @@
+/*
+	Copyright (C) 2006 - 2022
+	by Dominic Bolin <dominic.bolin@exong.net>
+	Part of the Battle for Wesnoth Project https://www.wesnoth.org/
+
+	This program is free software; you can redistribute it and/or modify
+	it under the terms of the GNU General Public License as published by
+	the Free Software Foundation; either version 2 of the License, or
+	(at your option) any later version.
+	This program is distributed in the hope that it will be useful,
+	but WITHOUT ANY WARRANTY.
+
+	See the COPYING file for more details.
+*/
+
+/**
+ *  @file
+ *  Manage unit-abilities, like heal, cure, and weapon_specials.
+ */
+
+#include "display.hpp"
+#include "display_context.hpp"
+#include "game_board.hpp"
+#include "gettext.hpp"
+#include "global.hpp"
+#include "lexical_cast.hpp"
+#include "log.hpp"
+#include "map/map.hpp"
+#include "resources.hpp"
+#include "units/unit.hpp"
+#include "units/active_ability_list.hpp"
+#include "units/map.hpp"
+#include "formula/callable_objects.hpp"
+#include "formula/formula.hpp"
+#include "formula/function_gamestate.hpp"
+
+#include <boost/dynamic_bitset.hpp>
+
+#include <string_view>
+
+static lg::log_domain log_engine("engine");
+#define ERR_NG LOG_STREAM(err, log_engine)
+
+static lg::log_domain log_wml("wml");
+#define ERR_WML LOG_STREAM(err, log_wml)
+
+namespace {
+
+const unit_map& get_unit_map()
+{
+	// Used if we're in the game, including during the construction of the display_context
+	if(resources::gameboard) {
+		return resources::gameboard->units();
+	}
+
+	// If we get here, we're in the scenario editor
+	assert(display::get_singleton());
+	return display::get_singleton()->get_units();
+}
+
+template<typename T, typename TFuncFormula>
+class get_ability_value_visitor
+#ifdef USING_BOOST_VARIANT
+	: public boost::static_visitor<T>
+#endif
+{
+public:
+	// Constructor stores the default value.
+	get_ability_value_visitor(T def, const TFuncFormula& formula_handler) : def_(def), formula_handler_(formula_handler) {}
+
+	T operator()(const utils::monostate&) const { return def_; }
+	T operator()(bool)                 const { return def_; }
+	T operator()(int i)                const { return static_cast<T>(i); }
+	T operator()(unsigned long long u) const { return static_cast<T>(u); }
+	T operator()(double d)             const { return static_cast<T>(d); }
+	T operator()(const t_string&)     const { return def_; }
+	T operator()(const std::string& s) const
+	{
+		if(s.size() >= 2 && s[0] == '(') {
+			return formula_handler_(s);
+		}
+		return lexical_cast_default<T>(s, def_);
+	}
+
+private:
+	const T def_;
+	const TFuncFormula& formula_handler_;
+};
+
+template<typename T, typename TFuncFormula>
+T get_single_ability_value(const config::attribute_value& v, T def, const active_ability& ability_info, const map_location& receiver_loc, const_attack_ptr att, const TFuncFormula& formula_handler)
+{
+	return v.apply_visitor(get_ability_value_visitor(def, [&](const std::string& s) {
+
+			try {
+				const unit_map& units = get_unit_map();
+
+				auto u_itor = units.find(ability_info.teacher_loc);
+
+				if(u_itor == units.end()) {
+					return def;
+				}
+				wfl::map_formula_callable callable(std::make_shared<wfl::unit_callable>(*u_itor));
+				if(att) {
+					att->add_formula_context(callable);
+				}
+				if (auto uptr = units.find_unit_ptr(ability_info.student_loc)) {
+					callable.add("student", wfl::variant(std::make_shared<wfl::unit_callable>(*uptr)));
+				}
+				if (auto uptr = units.find_unit_ptr(receiver_loc)) {
+					callable.add("other", wfl::variant(std::make_shared<wfl::unit_callable>(*uptr)));
+				}
+				return formula_handler(wfl::formula(s, new wfl::gamestate_function_symbol_table), callable);
+			} catch(const wfl::formula_error& e) {
+				lg::log_to_chat() << "Formula error in ability or weapon special: " << e.type << " at " << e.filename << ':' << e.line << ")\n";
+				ERR_WML << "Formula error in ability or weapon special: " << e.type << " at " << e.filename << ':' << e.line << ")";
+				return def;
+			}
+	}));
+}
+}
+
+template<typename TComp>
+std::pair<int,map_location> active_ability_list::get_extremum(const std::string& key, int def, const TComp& comp) const
+{
+	if ( cfgs_.empty() ) {
+		return std::pair(def, map_location());
+	}
+	// The returned location is the best non-cumulative one, if any,
+	// the best absolute cumulative one otherwise.
+	map_location best_loc;
+	bool only_cumulative = true;
+	int abs_max = 0;
+	int flat = 0;
+	int stack = 0;
+	for (const active_ability& p : cfgs_)
+	{
+		int value = get_single_ability_value(p.ability_cfg()[key], def, p, loc(), const_attack_ptr(), [&](const wfl::formula& formula, wfl::map_formula_callable& callable) {
+			return formula.evaluate(callable).as_int();
+		});
+
+		if (p.ability_cfg()["cumulative"].to_bool()) {
+			stack += value;
+			if (value < 0) value = -value;
+			if (only_cumulative && !comp(value, abs_max)) {
+				abs_max = value;
+				best_loc = p.teacher_loc;
+			}
+		} else if (only_cumulative || comp(flat, value)) {
+			only_cumulative = false;
+			flat = value;
+			best_loc = p.teacher_loc;
+		}
+	}
+	return std::pair(flat + stack, best_loc);
+}
+
+template std::pair<int, map_location> active_ability_list::get_extremum<std::less<int>>(const std::string& key, int def, const std::less<int>& comp) const;
+template std::pair<int, map_location> active_ability_list::get_extremum<std::greater<int>>(const std::string& key, int def, const std::greater<int>& comp) const;
+
+/*
+ *
+ * [special]
+ * [swarm]
+ *	name= _ "swarm"
+ *	name_inactive= _ ""
+ *	description= _ ""
+ *	description_inactive= _ ""
+ *	cumulative=no
+ *	apply_to=self  #self,opponent,defender,attacker,both
+ *	#active_on=defense # or offense; omitting this means "both"
+ *
+ *	swarm_attacks_max=4
+ *	swarm_attacks_min=2
+ *
+ *	[filter_self] // SUF
+ *		...
+ *	[/filter_self]
+ *	[filter_opponent] // SUF
+ *	[filter_attacker] // SUF
+ *	[filter_defender] // SUF
+ *	[filter_adjacent] // SAUF
+ *	[filter_adjacent_location] // SAUF + locs
+ * [/swarm]
+ * [/special]
+ *
+ */
+
+
+namespace unit_abilities
+{
+
+void individual_effect::set(value_modifier t, int val, const config *abil, const map_location &l)
+{
+	type=t;
+	value=val;
+	ability=abil;
+	loc=l;
+}
+
+bool filter_base_matches(const config& cfg, int def)
+{
+	if (auto apply_filter = cfg.optional_child("filter_base_value")) {
+		config::attribute_value cond_eq = apply_filter["equals"];
+		config::attribute_value cond_ne = apply_filter["not_equals"];
+		config::attribute_value cond_lt = apply_filter["less_than"];
+		config::attribute_value cond_gt = apply_filter["greater_than"];
+		config::attribute_value cond_ge = apply_filter["greater_than_equal_to"];
+		config::attribute_value cond_le = apply_filter["less_than_equal_to"];
+		return  (cond_eq.empty() || def == cond_eq.to_int()) &&
+			(cond_ne.empty() || def != cond_ne.to_int()) &&
+			(cond_lt.empty() || def <  cond_lt.to_int()) &&
+			(cond_gt.empty() || def >  cond_gt.to_int()) &&
+			(cond_ge.empty() || def >= cond_ge.to_int()) &&
+			(cond_le.empty() || def <= cond_le.to_int());
+	}
+	return true;
+}
+
+effect::effect(const active_ability_list& list, int def, const_attack_ptr att, bool is_cumulable) :
+	effect_list_(),
+	composite_value_(0)
+{
+
+	int value_set = is_cumulable ? std::max(list.highest("value").first, 0) + std::min(list.lowest("value").first, 0) : def;
+	std::map<std::string,individual_effect> values_add;
+	std::map<std::string,individual_effect> values_mul;
+	std::map<std::string,individual_effect> values_div;
+
+	individual_effect set_effect_max;
+	individual_effect set_effect_min;
+
+	for (const active_ability & ability : list) {
+		const config& cfg = ability.ability_cfg();
+		const std::string& effect_id = cfg[cfg["id"].empty() ? "name" : "id"];
+
+		if (!filter_base_matches(cfg, def))
+			continue;
+
+		if(!is_cumulable){
+			if (const config::attribute_value *v = cfg.get("value")) {
+				int value = get_single_ability_value(*v, def, ability, list.loc(), att, [&](const wfl::formula& formula, wfl::map_formula_callable& callable) {
+					callable.add("base_value", wfl::variant(def));
+					return formula.evaluate(callable).as_int();
+				});
+
+				int value_cum = cfg["cumulative"].to_bool() ? std::max(def, value) : value;
+				assert((set_effect_min.type != NOT_USED) == (set_effect_max.type != NOT_USED));
+				if(set_effect_min.type == NOT_USED) {
+					set_effect_min.set(SET, value_cum, &ability.ability_cfg(), ability.teacher_loc);
+					set_effect_max.set(SET, value_cum, &ability.ability_cfg(), ability.teacher_loc);
+				}
+				else {
+					if(value_cum > set_effect_max.value) {
+						set_effect_max.set(SET, value_cum, &ability.ability_cfg(), ability.teacher_loc);
+					}
+					if(value_cum < set_effect_min.value) {
+						set_effect_min.set(SET, value_cum, &ability.ability_cfg(), ability.teacher_loc);
+					}
+				}
+			}
+		}
+
+		if (const config::attribute_value *v = cfg.get("add")) {
+			int add = get_single_ability_value(*v, def, ability, list.loc(), att, [&](const wfl::formula& formula, wfl::map_formula_callable& callable) {
+				callable.add("base_value", wfl::variant(def));
+				return formula.evaluate(callable).as_int();
+			});
+			std::map<std::string,individual_effect>::iterator add_effect = values_add.find(effect_id);
+			if(add_effect == values_add.end() || add > add_effect->second.value) {
+				values_add[effect_id].set(ADD, add, &ability.ability_cfg(), ability.teacher_loc);
+			}
+		}
+		if (const config::attribute_value *v = cfg.get("sub")) {
+			int sub = - get_single_ability_value(*v, def, ability, list.loc(), att, [&](const wfl::formula& formula, wfl::map_formula_callable& callable) {
+				callable.add("base_value", wfl::variant(def));
+				return formula.evaluate(callable).as_int();
+			});
+			std::map<std::string,individual_effect>::iterator sub_effect = values_add.find(effect_id);
+			if(sub_effect == values_add.end() || sub < sub_effect->second.value) {
+				values_add[effect_id].set(ADD, sub, &ability.ability_cfg(), ability.teacher_loc);
+			}
+		}
+		if (const config::attribute_value *v = cfg.get("multiply")) {
+			int multiply = static_cast<int>(get_single_ability_value(*v, static_cast<double>(def), ability, list.loc(), att, [&](const wfl::formula& formula, wfl::map_formula_callable& callable) {
+				callable.add("base_value", wfl::variant(def));
+				return formula.evaluate(callable).as_decimal() / 1000.0 ;
+			}) * 100);
+			std::map<std::string,individual_effect>::iterator mul_effect = values_mul.find(effect_id);
+			if(mul_effect == values_mul.end() || multiply > mul_effect->second.value) {
+				values_mul[effect_id].set(MUL, multiply, &ability.ability_cfg(), ability.teacher_loc);
+			}
+		}
+		if (const config::attribute_value *v = cfg.get("divide")) {
+			int divide = static_cast<int>(get_single_ability_value(*v, static_cast<double>(def), ability, list.loc(), att, [&](const wfl::formula& formula, wfl::map_formula_callable& callable) {
+				callable.add("base_value", wfl::variant(def));
+				return formula.evaluate(callable).as_decimal() / 1000.0 ;
+			}) * 100);
+
+			if (divide == 0) {
+				ERR_NG << "division by zero with divide= in ability/weapon special " << effect_id;
+			}
+			else {
+				std::map<std::string,individual_effect>::iterator div_effect = values_div.find(effect_id);
+				if(div_effect == values_div.end() || divide > div_effect->second.value) {
+					values_div[effect_id].set(DIV, divide, &ability.ability_cfg(), ability.teacher_loc);
+				}
+			}
+		}
+	}
+
+	if(!is_cumulable && set_effect_max.type != NOT_USED) {
+		value_set = std::max(set_effect_max.value, 0) + std::min(set_effect_min.value, 0);
+		if(set_effect_max.value > def) {
+			effect_list_.push_back(set_effect_max);
+		}
+		if(set_effect_min.value < def) {
+			effect_list_.push_back(set_effect_min);
+		}
+	}
+
+	/* Do multiplication with floating point values rather than integers
+	 * We want two places of precision for each multiplier
+	 * Using integers multiplied by 100 to keep precision causes overflow
+	 *   after 3-4 abilities for 32-bit values and ~8 for 64-bit
+	 * Avoiding the overflow by dividing after each step introduces rounding errors
+	 *   that may vary depending on the order effects are applied
+	 * As the final values are likely <1000 (always true for mainline), loss of less significant digits is not an issue
+	 */
+	double multiplier = 1.0;
+	double divisor = 1.0;
+
+	for(const auto& val : values_mul) {
+		multiplier *= val.second.value/100.0;
+		effect_list_.push_back(val.second);
+	}
+
+	for(const auto& val : values_div) {
+		divisor *= val.second.value/100.0;
+		effect_list_.push_back(val.second);
+	}
+
+	int addition = 0;
+	for(const auto& val : values_add) {
+		addition += val.second.value;
+		effect_list_.push_back(val.second);
+	}
+
+	composite_value_ = static_cast<int>((value_set + addition) * multiplier / divisor);
+}
+
+} // end namespace unit_abilities
